Renaming mmf_routines to modules and placing them in a namespace like I should've done in the first place
